chore!: remove classic tuniform 2^-64 params from the lib
- vendor them in the bench crate to keep the current benchmark parameters
available

BREAKING CHANGE:
the PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64 parameter aliases has been
deprecated
