+++ tfhe-benchmark/src/params/shortint_params/classic/mod.rs
@@ -1,0 +1,32 @@
+use tfhe::shortint::parameters::v1_6;
+
+// Classic TUniform 2^-64 removed in 1.7
+pub use v1_6::classic::tuniform::p_fail_2_minus_64::ks_pbs::{
+    V1_6_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M64 as BENCH_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M64,
+    V1_6_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64 as BENCH_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64,
+    V1_6_PARAM_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M64 as BENCH_PARAM_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M64,
+    V1_6_PARAM_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M64 as BENCH_PARAM_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M64,
+};
+
+use tfhe::shortint::parameters::ClassicPBSParameters;
+
+// "Vendored" parameters which used to be in TFHE-rs but we want to keep around for now for
+// benchmarking
+pub const DEDICATED_BENCH_SHORTINT_CLASSIC_PBS_PARAMETERS: [(&ClassicPBSParameters, &str); 4] = [
+    (
+        &BENCH_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M64,
+        "BENCH_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M64",
+    ),
+    (
+        &BENCH_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64,
+        "BENCH_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64",
+    ),
+    (
+        &BENCH_PARAM_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M64,
+        "BENCH_PARAM_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M64",
+    ),
+    (
+        &BENCH_PARAM_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M64,
+        "BENCH_PARAM_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M64",
+    ),
+];
