+++ tfhe-benchmark/src/params/shortint_params/mod.rs
@@ -1,0 +1,551 @@
+pub mod classic;
+
+use crate::params_aliases::*;
+use std::collections::HashMap;
+use std::env;
+use std::sync::OnceLock;
+use tfhe::core_crypto::prelude::{DynamicDistribution, LweBskGroupingFactor};
+use tfhe::shortint::{
+    AtomicPatternParameters, CarryModulus, ClassicPBSParameters, MessageModulus,
+    MultiBitPBSParameters,
+};
+
+pub const SHORTINT_BENCH_PARAMS_TUNIFORM: [(ClassicPBSParameters, &str); 4] = [
+    (
+        BENCH_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M128,
+        "BENCH_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M128",
+    ),
+    (
+        BENCH_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M128,
+        "BENCH_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M128",
+    ),
+    (
+        BENCH_PARAM_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M128,
+        "BENCH_PARAM_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M128",
+    ),
+    (
+        BENCH_PARAM_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M128,
+        "BENCH_PARAM_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M128",
+    ),
+];
+
+pub const SHORTINT_BENCH_PARAMS_GAUSSIAN: [(ClassicPBSParameters, &str); 4] = [
+    (
+        BENCH_PARAM_MESSAGE_1_CARRY_1_KS_PBS_GAUSSIAN_2M128,
+        "BENCH_PARAM_MESSAGE_1_CARRY_1_KS_PBS_GAUSSIAN_2M128",
+    ),
+    (
+        BENCH_PARAM_MESSAGE_2_CARRY_2_KS_PBS_GAUSSIAN_2M128,
+        "BENCH_PARAM_MESSAGE_2_CARRY_2_KS_PBS_GAUSSIAN_2M128",
+    ),
+    (
+        BENCH_PARAM_MESSAGE_3_CARRY_3_KS_PBS_GAUSSIAN_2M128,
+        "BENCH_PARAM_MESSAGE_3_CARRY_3_KS_PBS_GAUSSIAN_2M128",
+    ),
+    (
+        BENCH_PARAM_MESSAGE_4_CARRY_4_KS_PBS_GAUSSIAN_2M128,
+        "BENCH_PARAM_MESSAGE_4_CARRY_4_KS_PBS_GAUSSIAN_2M128",
+    ),
+];
+
+pub const SHORTINT_BENCH_PARAMS_TUNIFORM_DOCUMENTATION: [(ClassicPBSParameters, &str); 8] = [
+    (
+        BENCH_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M64,
+        "BENCH_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M64,",
+    ),
+    (
+        BENCH_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64,
+        "BENCH_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64,",
+    ),
+    (
+        BENCH_PARAM_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M64,
+        "BENCH_PARAM_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M64,",
+    ),
+    (
+        BENCH_PARAM_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M64,
+        "BENCH_PARAM_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M64,",
+    ),
+    (
+        BENCH_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M128,
+        "BENCH_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M128,",
+    ),
+    (
+        BENCH_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M128,
+        "BENCH_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M128,",
+    ),
+    (
+        BENCH_PARAM_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M128,
+        "BENCH_PARAM_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M128,",
+    ),
+    (
+        BENCH_PARAM_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M128,
+        "BENCH_PARAM_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M128,",
+    ),
+];
+
+pub const SHORTINT_MULTI_BIT_BENCH_PARAMS: &[(MultiBitPBSParameters, &str)] = {
+    #[cfg(feature = "gpu")]
+    {
+        &[
+            // TUniform
+            (
+                BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M128,
+                "BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M128",
+            ),
+            (
+                BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M128,
+                "BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M128",
+            ),
+            (
+                BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M128,
+                "BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M128",
+            ),
+            // Gaussian
+            (
+                BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_1_CARRY_1_KS_PBS_GAUSSIAN_2M128,
+                "BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_1_CARRY_1_KS_PBS_GAUSSIAN_2M128",
+            ),
+            (
+                BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_2_CARRY_2_KS_PBS_GAUSSIAN_2M128,
+                "BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_2_CARRY_2_KS_PBS_GAUSSIAN_2M128",
+            ),
+            (
+                BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_3_CARRY_3_KS_PBS_GAUSSIAN_2M128,
+                "BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_3_CARRY_3_KS_PBS_GAUSSIAN_2M128",
+            ),
+        ]
+    }
+
+    #[cfg(not(feature = "gpu"))]
+    {
+        &[
+            // Group 2
+            (
+                BENCH_PARAM_MULTI_BIT_GROUP_2_MESSAGE_1_CARRY_1_KS_PBS_GAUSSIAN_2M128,
+                "BENCH_PARAM_MULTI_BIT_GROUP_2_MESSAGE_1_CARRY_1_KS_PBS_GAUSSIAN_2M128",
+            ),
+            (
+                BENCH_PARAM_MULTI_BIT_GROUP_2_MESSAGE_2_CARRY_2_KS_PBS_GAUSSIAN_2M128,
+                "BENCH_PARAM_MULTI_BIT_GROUP_2_MESSAGE_2_CARRY_2_KS_PBS_GAUSSIAN_2M128",
+            ),
+            (
+                BENCH_PARAM_MULTI_BIT_GROUP_2_MESSAGE_3_CARRY_3_KS_PBS_GAUSSIAN_2M128,
+                "BENCH_PARAM_MULTI_BIT_GROUP_2_MESSAGE_3_CARRY_3_KS_PBS_GAUSSIAN_2M128",
+            ),
+            // Group 3
+            (
+                BENCH_PARAM_MULTI_BIT_GROUP_3_MESSAGE_1_CARRY_1_KS_PBS_GAUSSIAN_2M128,
+                "BENCH_PARAM_MULTI_BIT_GROUP_3_MESSAGE_1_CARRY_1_KS_PBS_GAUSSIAN_2M128",
+            ),
+            (
+                BENCH_PARAM_MULTI_BIT_GROUP_3_MESSAGE_2_CARRY_2_KS_PBS_GAUSSIAN_2M128,
+                "BENCH_PARAM_MULTI_BIT_GROUP_3_MESSAGE_2_CARRY_2_KS_PBS_GAUSSIAN_2M128",
+            ),
+            (
+                BENCH_PARAM_MULTI_BIT_GROUP_3_MESSAGE_3_CARRY_3_KS_PBS_GAUSSIAN_2M128,
+                "BENCH_PARAM_MULTI_BIT_GROUP_3_MESSAGE_3_CARRY_3_KS_PBS_GAUSSIAN_2M128",
+            ),
+        ]
+    }
+};
+
+pub const SHORTINT_MULTI_BIT_BENCH_PARAMS_DOCUMENTATION: &[(MultiBitPBSParameters, &str)] = {
+    #[cfg(feature = "gpu")]
+    {
+        &[
+            // 2^-64
+            (
+                BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M64,
+                "BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M64",
+            ),
+            (
+                BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64,
+                "BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64",
+            ),
+            (
+                BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M64,
+                "BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M64",
+            ),
+            // 2^-128
+            (
+                BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M128,
+                "BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M128",
+            ),
+            (
+                BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M128,
+                "BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M128",
+            ),
+            (
+                BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M128,
+                "BENCH_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M128",
+            ),
+        ]
+    }
+
+    #[cfg(not(feature = "gpu"))]
+    {
+        &[
+            // 2^-64
+            (
+                BENCH_PARAM_MULTI_BIT_GROUP_4_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M64,
+                "BENCH_PARAM_MULTI_BIT_GROUP_4_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M64",
+            ),
+            (
+                BENCH_PARAM_MULTI_BIT_GROUP_4_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64,
+                "BENCH_PARAM_MULTI_BIT_GROUP_4_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64",
+            ),
+            (
+                BENCH_PARAM_MULTI_BIT_GROUP_4_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M64,
+                "BENCH_PARAM_MULTI_BIT_GROUP_4_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M64",
+            ),
+            (
+                BENCH_PARAM_MULTI_BIT_GROUP_4_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M64,
+                "BENCH_PARAM_MULTI_BIT_GROUP_4_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M64",
+            ),
+            // 2^-128
+            (
+                BENCH_PARAM_MULTI_BIT_GROUP_4_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M128,
+                "BENCH_PARAM_MULTI_BIT_GROUP_4_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M128",
+            ),
+            (
+                BENCH_PARAM_MULTI_BIT_GROUP_4_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M128,
+                "BENCH_PARAM_MULTI_BIT_GROUP_4_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M128",
+            ),
+            (
+                BENCH_PARAM_MULTI_BIT_GROUP_4_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M128,
+                "BENCH_PARAM_MULTI_BIT_GROUP_4_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M128",
+            ),
+            (
+                BENCH_PARAM_MULTI_BIT_GROUP_4_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M128,
+                "BENCH_PARAM_MULTI_BIT_GROUP_4_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M128",
+            ),
+        ]
+    }
+};
+
+use tfhe::shortint::parameters::MetaParameters;
+
+#[cfg(feature = "internal-keycache")]
+pub mod shortint_params_keycache {
+    use super::*;
+    use crate::crypto_record::{BenchPackingKsParams, BenchPbsParams};
+    use crate::utilities::{get_param_type, ParamType};
+
+    pub fn benchmark_parameters() -> Vec<(String, BenchPbsParams<u64>)> {
+        match get_parameters_set() {
+            ParametersSet::Default => {
+                let iterator = match get_param_type() {
+                    ParamType::ClassicalDocumentation => {
+                        SHORTINT_BENCH_PARAMS_TUNIFORM_DOCUMENTATION
+                            .iter()
+                            .chain([].iter())
+                    }
+                    _ => SHORTINT_BENCH_PARAMS_TUNIFORM
+                        .iter()
+                        .chain(SHORTINT_BENCH_PARAMS_GAUSSIAN.iter()),
+                };
+                iterator
+                    .map(|(params, name)| (name.to_string(), (*params).into()))
+                    .collect()
+            }
+            ParametersSet::All => filter_parameters(
+                BENCH_ALL_CLASSIC_PBS_PARAMETERS.into_iter().flatten(),
+                DesiredNoiseDistribution::Both,
+                DesiredBackend::Cpu,
+            )
+            .into_iter()
+            .map(|(params, name)| (name.to_string(), (*params).into()))
+            .collect(),
+        }
+    }
+
+    pub fn benchmark_compression_parameters() -> Vec<(String, BenchPackingKsParams<u64>)> {
+        use tfhe::keycache::NamedParam;
+        vec![(
+            BENCH_COMP_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M128.name(),
+            (
+                BENCH_COMP_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M128,
+                BENCH_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M128.into(),
+            )
+                .into(),
+        )]
+    }
+
+    pub fn multi_bit_benchmark_parameters(
+    ) -> Vec<(String, BenchPbsParams<u64>, LweBskGroupingFactor)> {
+        match get_parameters_set() {
+            ParametersSet::Default => match get_param_type() {
+                ParamType::MultiBitDocumentation => SHORTINT_MULTI_BIT_BENCH_PARAMS_DOCUMENTATION
+                    .iter()
+                    .map(|(params, name)| {
+                        (name.to_string(), (*params).into(), params.grouping_factor)
+                    })
+                    .collect(),
+                _ => SHORTINT_MULTI_BIT_BENCH_PARAMS
+                    .iter()
+                    .map(|(params, name)| {
+                        (name.to_string(), (*params).into(), params.grouping_factor)
+                    })
+                    .collect(),
+            },
+            ParametersSet::All => {
+                let desired_backend = if cfg!(feature = "gpu") {
+                    DesiredBackend::Gpu
+                } else {
+                    DesiredBackend::Cpu
+                };
+                filter_parameters(
+                    &BENCH_ALL_MULTI_BIT_PBS_PARAMETERS,
+                    DesiredNoiseDistribution::Both,
+                    desired_backend,
+                )
+                .into_iter()
+                .map(|(params, name)| (name.to_string(), (*params).into(), params.grouping_factor))
+                .collect()
+            }
+        }
+    }
+}
+
+#[cfg(feature = "internal-keycache")]
+pub use shortint_params_keycache::*;
+
+pub fn raw_benchmark_parameters() -> Vec<(String, AtomicPatternParameters)> {
+    let is_multi_bit = match env::var("__TFHE_RS_PARAM_TYPE") {
+        Ok(val) => val.to_lowercase() == "multi_bit",
+        Err(_) => false,
+    };
+
+    if is_multi_bit {
+        SHORTINT_MULTI_BIT_BENCH_PARAMS
+            .iter()
+            .map(|(p, name)| (name.to_string(), (*p).into()))
+            .collect()
+    } else {
+        SHORTINT_BENCH_PARAMS_TUNIFORM
+            .iter()
+            .chain(SHORTINT_BENCH_PARAMS_GAUSSIAN.iter())
+            .map(|(p, name)| (name.to_string(), (*p).into()))
+            .collect()
+    }
+}
+
+// This array has been built according to performance benchmarks measuring latency over a
+// matrix of 4 parameters set, 3 grouping factor and a wide range of threads values.
+// The values available here as u64 are the optimal number of threads to use for a given triplet
+// representing one or more parameters set.
+const MULTI_BIT_THREADS_ARRAY: [((MessageModulus, CarryModulus, LweBskGroupingFactor), u64); 12] = [
+    (
+        (MessageModulus(2), CarryModulus(2), LweBskGroupingFactor(2)),
+        5,
+    ),
+    (
+        (MessageModulus(4), CarryModulus(4), LweBskGroupingFactor(2)),
+        5,
+    ),
+    (
+        (MessageModulus(8), CarryModulus(8), LweBskGroupingFactor(2)),
+        5,
+    ),
+    (
+        (
+            MessageModulus(16),
+            CarryModulus(16),
+            LweBskGroupingFactor(2),
+        ),
+        5,
+    ),
+    (
+        (MessageModulus(2), CarryModulus(2), LweBskGroupingFactor(3)),
+        7,
+    ),
+    (
+        (MessageModulus(4), CarryModulus(4), LweBskGroupingFactor(3)),
+        9,
+    ),
+    (
+        (MessageModulus(8), CarryModulus(8), LweBskGroupingFactor(3)),
+        10,
+    ),
+    (
+        (
+            MessageModulus(16),
+            CarryModulus(16),
+            LweBskGroupingFactor(3),
+        ),
+        10,
+    ),
+    (
+        (MessageModulus(2), CarryModulus(2), LweBskGroupingFactor(4)),
+        11,
+    ),
+    (
+        (MessageModulus(4), CarryModulus(4), LweBskGroupingFactor(4)),
+        13,
+    ),
+    (
+        (MessageModulus(8), CarryModulus(8), LweBskGroupingFactor(4)),
+        11,
+    ),
+    (
+        (
+            MessageModulus(16),
+            CarryModulus(16),
+            LweBskGroupingFactor(4),
+        ),
+        11,
+    ),
+];
+
+/// Define the number of threads to use for  parameters doing multithreaded programmable
+/// bootstrapping.
+///
+/// Parameters must have the same values between message and carry modulus.
+/// Grouping factor 2, 3 and 4 are the only ones that are supported.
+pub fn multi_bit_num_threads(
+    message_modulus: u64,
+    carry_modulus: u64,
+    grouping_factor: usize,
+) -> Option<u64> {
+    // TODO Implement an interpolation mechanism for X_Y parameters set
+    if message_modulus != carry_modulus || ![2, 3, 4].contains(&(grouping_factor as i32)) {
+        return None;
+    }
+    let thread_map: HashMap<(MessageModulus, CarryModulus, LweBskGroupingFactor), u64> =
+        HashMap::from_iter(MULTI_BIT_THREADS_ARRAY);
+    thread_map
+        .get(&(
+            MessageModulus(message_modulus),
+            CarryModulus(carry_modulus),
+            LweBskGroupingFactor(grouping_factor),
+        ))
+        .copied()
+}
+
+pub static PARAMETERS_SET: OnceLock<ParametersSet> = OnceLock::new();
+
+pub enum ParametersSet {
+    Default,
+    All,
+}
+
+impl ParametersSet {
+    pub fn from_env() -> Result<Self, String> {
+        let raw_value = env::var("__TFHE_RS_PARAMS_SET").unwrap_or("default".to_string());
+        match raw_value.to_lowercase().as_str() {
+            "default" => Ok(ParametersSet::Default),
+            "all" => Ok(ParametersSet::All),
+            _ => Err(format!("parameters set '{raw_value}' is not supported")),
+        }
+    }
+}
+
+pub fn get_parameters_set() -> &'static ParametersSet {
+    PARAMETERS_SET.get_or_init(|| ParametersSet::from_env().unwrap())
+}
+
+#[derive(Clone, Copy, Debug)]
+pub enum DesiredNoiseDistribution {
+    Gaussian,
+    TUniform,
+    Both,
+}
+
+#[derive(Clone, Copy, Debug)]
+pub enum DesiredBackend {
+    Cpu,
+    Gpu,
+}
+
+impl DesiredBackend {
+    fn matches_parameter_name_backend(&self, param_name: &str) -> bool {
+        matches!(
+            (self, param_name.to_lowercase().contains("gpu")),
+            (DesiredBackend::Cpu, false) | (DesiredBackend::Gpu, true)
+        )
+    }
+}
+
+pub fn filter_parameters<'a, P, PIter>(
+    params: PIter,
+    desired_noise_distribution: DesiredNoiseDistribution,
+    desired_backend: DesiredBackend,
+) -> Vec<(&'a P, &'a str)>
+where
+    P: Copy + Into<AtomicPatternParameters>,
+    PIter: IntoIterator<Item = &'a (&'a P, &'a str)>,
+{
+    params
+        .into_iter()
+        .filter_map(|(p, name)| {
+            let temp_param: AtomicPatternParameters = (**p).into();
+
+            match (
+                temp_param.lwe_noise_distribution(),
+                desired_noise_distribution,
+            ) {
+                // If it's one of the pairs, we continue the process.
+                (DynamicDistribution::Gaussian(_), DesiredNoiseDistribution::Gaussian)
+                | (DynamicDistribution::TUniform(_), DesiredNoiseDistribution::TUniform)
+                | (_, DesiredNoiseDistribution::Both) => (),
+                _ => return None,
+            }
+
+            if !desired_backend.matches_parameter_name_backend(name) {
+                return None;
+            };
+
+            Some((*p, *name))
+        })
+        .collect()
+}
+
+pub fn get_classical_tuniform_groups() -> Vec<MetaParameters> {
+    vec![
+        // Most complete 2_2 parameters set
+        BENCH_META_PARAM_CPU_2_2_KS_PBS_PKE_TO_SMALL_ZKV2_TUNIFORM_2M128,
+        BENCH_META_PARAM_CPU_2_2_KS32_PBS_PKE_TO_SMALL_ZKV2_TUNIFORM_2M128,
+        BENCH_META_PARAM_CPU_1_1_KS_PBS_TUNIFORM_2M128,
+        BENCH_META_PARAM_CPU_3_3_KS_PBS_TUNIFORM_2M128,
+        BENCH_META_PARAM_CPU_4_4_KS_PBS_TUNIFORM_2M128,
+    ]
+}
+
+pub fn get_multi_bit_tuniform_groups() -> Vec<MetaParameters> {
+    // TODO Complete parameters groups once corresponding parameters set are available.
+    vec![
+        // Group 2
+        // CPU ---
+        BENCH_META_PARAM_CPU_1_1_MULTI_BIT_GROUP_2_KS_PBS_TUNIFORM_2M128,
+        BENCH_META_PARAM_CPU_2_2_MULTI_BIT_GROUP_2_KS_PBS_TUNIFORM_2M128,
+        BENCH_META_PARAM_CPU_3_3_MULTI_BIT_GROUP_2_KS_PBS_TUNIFORM_2M128,
+        BENCH_META_PARAM_CPU_4_4_MULTI_BIT_GROUP_2_KS_PBS_TUNIFORM_2M128,
+        // GPU ---
+        BENCH_META_PARAM_GPU_1_1_MULTI_BIT_GROUP_2_KS_PBS_GAUSSIAN_2M128,
+        BENCH_META_PARAM_GPU_2_2_MULTI_BIT_GROUP_2_KS_PBS_GAUSSIAN_2M128,
+        BENCH_META_PARAM_GPU_3_3_MULTI_BIT_GROUP_2_KS_PBS_GAUSSIAN_2M128,
+        BENCH_META_PARAM_GPU_4_4_MULTI_BIT_GROUP_2_KS_PBS_GAUSSIAN_2M128,
+        // Group 3
+        // CPU ---
+        BENCH_META_PARAM_CPU_1_1_MULTI_BIT_GROUP_3_KS_PBS_TUNIFORM_2M128,
+        BENCH_META_PARAM_CPU_2_2_MULTI_BIT_GROUP_3_KS_PBS_TUNIFORM_2M128,
+        BENCH_META_PARAM_CPU_3_3_MULTI_BIT_GROUP_3_KS_PBS_TUNIFORM_2M128,
+        BENCH_META_PARAM_CPU_4_4_MULTI_BIT_GROUP_3_KS_PBS_TUNIFORM_2M128,
+        // GPU ---
+        BENCH_META_PARAM_GPU_1_1_MULTI_BIT_GROUP_3_KS_PBS_GAUSSIAN_2M128,
+        BENCH_META_PARAM_GPU_2_2_MULTI_BIT_GROUP_3_KS_PBS_GAUSSIAN_2M128,
+        BENCH_META_PARAM_GPU_3_3_MULTI_BIT_GROUP_3_KS_PBS_GAUSSIAN_2M128,
+        BENCH_META_PARAM_GPU_4_4_MULTI_BIT_GROUP_3_KS_PBS_GAUSSIAN_2M128,
+        // Group 4
+        // CPU ---
+        BENCH_META_PARAM_CPU_1_1_MULTI_BIT_GROUP_4_KS_PBS_TUNIFORM_2M128,
+        BENCH_META_PARAM_CPU_2_2_MULTI_BIT_GROUP_4_KS_PBS_TUNIFORM_2M128,
+        BENCH_META_PARAM_CPU_3_3_MULTI_BIT_GROUP_4_KS_PBS_TUNIFORM_2M128,
+        BENCH_META_PARAM_CPU_4_4_MULTI_BIT_GROUP_4_KS_PBS_TUNIFORM_2M128,
+        // GPU ---
+        BENCH_META_PARAM_GPU_1_1_MULTI_BIT_GROUP_4_KS_PBS_GAUSSIAN_2M128,
+        BENCH_META_PARAM_GPU_2_2_MULTI_BIT_GROUP_4_KS_PBS_GAUSSIAN_2M128,
+        BENCH_META_PARAM_GPU_3_3_MULTI_BIT_GROUP_4_KS_PBS_GAUSSIAN_2M128,
+        BENCH_META_PARAM_GPU_4_4_MULTI_BIT_GROUP_4_KS_PBS_GAUSSIAN_2M128,
+    ]
+}
