+++ tfhe-benchmark/src/params_aliases.rs
@@ -25,14 +25,13 @@ pub mod shortint_params_aliases {
         V1_7_PARAM_MESSAGE_4_CARRY_4_KS_PBS_GAUSSIAN_2M128;
 
     // KS PBS TUniform
-    pub const BENCH_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M64: ClassicPBSParameters =
-        V1_7_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M64;
-    pub const BENCH_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64: ClassicPBSParameters =
-        V1_7_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64;
-    pub const BENCH_PARAM_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M64: ClassicPBSParameters =
-        V1_7_PARAM_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M64;
-    pub const BENCH_PARAM_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M64: ClassicPBSParameters =
-        V1_7_PARAM_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M64;
+    pub use crate::params::shortint_params::classic::{
+        BENCH_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M64,
+        BENCH_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64,
+        BENCH_PARAM_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M64,
+        BENCH_PARAM_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M64,
+    };
+
     #[cfg(not(feature = "gpu"))]
     pub const BENCH_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M128: ClassicPBSParameters =
         V1_7_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M128;
@@ -50,8 +49,11 @@ pub mod shortint_params_aliases {
     pub const BENCH_PARAM_MESSAGE_2_CARRY_2_KS32_PBS: KeySwitch32PBSParameters =
         V1_7_PARAM_MESSAGE_2_CARRY_2_KS32_PBS_TUNIFORM_2M128;
 
-    pub const BENCH_ALL_CLASSIC_PBS_PARAMETERS: [(&ClassicPBSParameters, &str); 54] =
-        VEC_ALL_CLASSIC_PBS_PARAMETERS;
+    // Manage the two sources of benchmark params we have: lib and dedicated bench params
+    pub const BENCH_ALL_CLASSIC_PBS_PARAMETERS: [&[(&ClassicPBSParameters, &str)]; 2] = [
+        &VEC_ALL_CLASSIC_PBS_PARAMETERS,
+        &crate::params::shortint_params::classic::DEDICATED_BENCH_SHORTINT_CLASSIC_PBS_PARAMETERS,
+    ];
 
     // MultiBit
     // CPU Gaussian
