+++ tfhe/src/shortint/keycache.rs
@@ -216,10 +216,6 @@ named_params_impl!( ShortintParameterSet =>
     V1_7_PARAM_GPU_MULTI_BIT_GROUP_4_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M128,
     // TUniform
     PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M128,
-    V1_7_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M64,
-    V1_7_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64,
-    V1_7_PARAM_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M64,
-    V1_7_PARAM_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M64,
     V1_7_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M128,
     V1_7_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M128,
     V1_7_PARAM_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M128,
