+++ tfhe/src/shortint/parameters/aliases.rs
@@ -10,7 +10,6 @@ use current_params::classic::gaussian::p_fail_2_minus_128::ks_pbs::{
 };
 use current_params::classic::gaussian::p_fail_2_minus_64::ks_pbs::V1_7_PARAM_MESSAGE_2_CARRY_2_KS_PBS_GAUSSIAN_2M64;
 use current_params::classic::tuniform::p_fail_2_minus_128::ks_pbs::V1_7_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M128;
-use current_params::classic::tuniform::p_fail_2_minus_64::ks_pbs::V1_7_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64;
 use current_params::compact_public_key_only::p_fail_2_minus_128::ks_pbs::V1_7_PARAM_PKE_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M128;
 use current_params::key_switching::p_fail_2_minus_128::ks_pbs::{
     V1_7_PARAM_KEYSWITCH_PKE_TO_BIG_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M128,
@@ -78,10 +77,6 @@ pub const PARAM_MESSAGE_2_CARRY_2_KS_PBS: ClassicPBSParameters =
     PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M128;
 
 pub const PARAM_MESSAGE_2_CARRY_2: ClassicPBSParameters = PARAM_MESSAGE_2_CARRY_2_KS_PBS;
-
-// 2M64
-pub const PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64: ClassicPBSParameters =
-    V1_7_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64;
 
 // Compute KS32
 // 2M128
