+++ tfhe/src/shortint/parameters/v1_7/classic/tuniform/mod.rs
@@ -1,2 +1,1 @@
 pub mod p_fail_2_minus_128;
-pub mod p_fail_2_minus_64;
