--- tfhe/src/shortint/parameters/v1_7/classic/tuniform/p_fail_2_minus_64/ks_pbs.rs
@@ -1,13 +1,0 @@
-use crate::shortint::parameters::ClassicPBSParameters;
-/// p-fail = 2^-72.226, algorithmic cost ~ 64, 2-norm = 3
-pub const V1_7_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M64: ClassicPBSParameters =
-    crate::shortint::parameters::v1_1::V1_1_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M64;
-/// p-fail = 2^-72.178, algorithmic cost ~ 112, 2-norm = 5
-pub const V1_7_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64: ClassicPBSParameters =
-    crate::shortint::parameters::v1_1::V1_1_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64;
-/// p-fail = 2^-64.454, algorithmic cost ~ 850, 2-norm = 9
-pub const V1_7_PARAM_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M64: ClassicPBSParameters =
-    crate::shortint::parameters::v1_1::V1_1_PARAM_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M64;
-/// p-fail = 2^-64.037, algorithmic cost ~ 6737, 2-norm = 17
-pub const V1_7_PARAM_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M64: ClassicPBSParameters =
-    crate::shortint::parameters::v1_1::V1_1_PARAM_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M64;
