--- tfhe/src/shortint/parameters/v1_7/classic/tuniform/p_fail_2_minus_64/mod.rs
@@ -1,1 +1,0 @@
-pub mod ks_pbs;
