+++ tfhe/src/shortint/parameters/v1_7/mod.rs
@@ -16,7 +16,6 @@ pub use classic::gaussian::p_fail_2_minus_128::pbs_ks::*;
 pub use classic::gaussian::p_fail_2_minus_64::ks_pbs::*;
 pub use classic::tuniform::p_fail_2_minus_128::ks_pbs::*;
 pub use classic::tuniform::p_fail_2_minus_128::ks_pbs_gpu::*;
-pub use classic::tuniform::p_fail_2_minus_64::ks_pbs::*;
 pub use compact_public_key_only::p_fail_2_minus_128::ks_pbs::*;
 pub use key_switching::p_fail_2_minus_128::ks_pbs::*;
 pub use ks32::tuniform::p_fail_2_minus_128::ks_pbs::*;
@@ -41,7 +40,7 @@ use crate::shortint::parameters::{
 };
 
 /// All [`ClassicPBSParameters`] in this module.
-pub const VEC_ALL_CLASSIC_PBS_PARAMETERS: [(&ClassicPBSParameters, &str); 54] = [
+pub const VEC_ALL_CLASSIC_PBS_PARAMETERS: [(&ClassicPBSParameters, &str); 50] = [
     (
         &V1_7_PARAM_MESSAGE_1_CARRY_1_KS_PBS_GAUSSIAN_2M64,
         "V1_7_PARAM_MESSAGE_1_CARRY_1_KS_PBS_GAUSSIAN_2M64",
@@ -217,22 +216,6 @@ pub const VEC_ALL_CLASSIC_PBS_PARAMETERS: [(&ClassicPBSParameters, &str); 54] = 
     (
         &V1_7_PARAM_MESSAGE_4_CARRY_4_PBS_KS_GAUSSIAN_2M128,
         "V1_7_PARAM_MESSAGE_4_CARRY_4_PBS_KS_GAUSSIAN_2M128",
-    ),
-    (
-        &V1_7_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M64,
-        "V1_7_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M64",
-    ),
-    (
-        &V1_7_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64,
-        "V1_7_PARAM_MESSAGE_2_CARRY_2_KS_PBS_TUNIFORM_2M64",
-    ),
-    (
-        &V1_7_PARAM_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M64,
-        "V1_7_PARAM_MESSAGE_3_CARRY_3_KS_PBS_TUNIFORM_2M64",
-    ),
-    (
-        &V1_7_PARAM_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M64,
-        "V1_7_PARAM_MESSAGE_4_CARRY_4_KS_PBS_TUNIFORM_2M64",
     ),
     (
         &V1_7_PARAM_MESSAGE_1_CARRY_1_KS_PBS_TUNIFORM_2M128,
@@ -1195,7 +1178,7 @@ pub const VEC_ALL_HPU_PARAMETERS: [(&KeySwitch32PBSParameters, &str); 3] = [
 ];
 
 /// All [`MetaParameters`] in this module.
-pub const VEC_ALL_META_PARAMETERS: [(&MetaParameters, &str); 119] = [
+pub const VEC_ALL_META_PARAMETERS: [(&MetaParameters, &str); 115] = [
     (
         &V1_7_META_PARAM_CPU_1_1_KS_PBS_GAUSSIAN_2M64,
         "V1_7_META_PARAM_CPU_1_1_KS_PBS_GAUSSIAN_2M64",
@@ -1227,22 +1210,6 @@ pub const VEC_ALL_META_PARAMETERS: [(&MetaParameters, &str); 119] = [
     (
         &V1_7_META_PARAM_CPU_4_4_KS_PBS_GAUSSIAN_2M128,
         "V1_7_META_PARAM_CPU_4_4_KS_PBS_GAUSSIAN_2M128",
-    ),
-    (
-        &V1_7_META_PARAM_CPU_1_1_KS_PBS_TUNIFORM_2M64,
-        "V1_7_META_PARAM_CPU_1_1_KS_PBS_TUNIFORM_2M64",
-    ),
-    (
-        &V1_7_META_PARAM_CPU_2_2_KS_PBS_TUNIFORM_2M64,
-        "V1_7_META_PARAM_CPU_2_2_KS_PBS_TUNIFORM_2M64",
-    ),
-    (
-        &V1_7_META_PARAM_CPU_3_3_KS_PBS_TUNIFORM_2M64,
-        "V1_7_META_PARAM_CPU_3_3_KS_PBS_TUNIFORM_2M64",
-    ),
-    (
-        &V1_7_META_PARAM_CPU_4_4_KS_PBS_TUNIFORM_2M64,
-        "V1_7_META_PARAM_CPU_4_4_KS_PBS_TUNIFORM_2M64",
     ),
     (
         &V1_7_META_PARAM_CPU_1_1_KS_PBS_TUNIFORM_2M128,
